+++ CHANGELOG.md
@@ -25,6 +25,12 @@
   - This change removes the need for `Sync` in types.
   - `'static` maybe appear as an additional requirment, but was previously implied.
   - `FnMut` is now used instead of `Fn` allowing state to be kept without the need of synchronization primatives.
+- **BREAKING** `InputHookFn` type alias has been removed.
+- **BREAKING** `InputHookRes` has been renamed to `InputHookCtrl`.
+  - **BREAKING** `Error` & `Warning` varients are now removed.
+    - Users should instead print the message themselves.
+  - **BREAKING** `Success` varient has been renamed to `Retain`.
+  - `InputHookCtrl` now implements `Default` and will default to `Retain`.
 - `AtlasImage` now has the `load_from_bytes`, `load_from_path`, `load_from_url` methods that are used by the corresponding `Atlas` methods.
 - `BstImageView` now the the `set_drop_fn` for setting a method to be called when all temporary views are dropped.
 - `BinPosition`, `BstEvent`, & `BstWinEv` types that already derived `PartialEq` now also derive `Eq`.